Sainsmart 2in1 Control Board
Rough mock-up, please verify current accuracy and continue to add detail if possible.  
Will make easier work of further iterations of covers, wire management, and various mods. 
In general it is easier(for me at least) do design if there is a frame of reference to start from,  build around, and terminate to.
If I have time tomorrow I will cad up a wall receptacle, and 3 prong plug and keep heading that direction.  Wish me good luck and godspeed on my journey to the nozzle.